# Replace Lists with arrays in SelectNeighbors()