For speed. A List has a fair amount of overhead from the extra palloc per element, as well as pointer indirection when walking the lists. This also allows us to use a faster inlined sort function.

I was expecting an up to 5%-10% speedup from this, based on profiling with perf. The sort was taking about 5% of CPU time, and lappend() and palloc() are 1%-2% each, so I figured that would be the maximum this could yield. But I'm actually seeing much more gain, up to 20%. So this is surprisingly effective - to the point that I wonder if I screwed up something. But it passes tests, and I haven't been able to spot a bug.