@jkatz 2023-12-23
Cool! I am going to kick off a benchmark on this vs. HEAD