@jkatz 2023-12-27
@hlinnaka Just a quick update -- I ran ANN Benchmarks on a fairly large machine comparing this to HEAD; I'm working on running some general indexing tests. Eyeballing the ANN Benchmark results, I saw a marginal gain on #388 that was more inline with your initial expectation. I'll continue running the tests and published a more detailed report.