@jkatz 2024-01-04
@hlinnaka What kind of tests were you running and what were the different parameters (vector size, build parameters, search parameters)? I'm not seeing the same performance uplift as you, but rather in the no-more-than 5% range.