@hlinnaka 2024-01-05
> @hlinnaka What kind of tests were you running and what were the different parameters (vector size, build parameters, search parameters)? I'm not seeing the same performance uplift as you, but rather in the no-more-than 5% range.

I just rebased this, and repeated the test I did earlier. This is what I did to initialize the test:
```
CREATE TABLE items_small (id int, embedding vector(100));
ALTER TABLE items_small ALTER COLUMN embedding SET STORAGE PLAIN;
-- `items.data` is the `glove-100-angular` dataset.
\COPY items_small FROM PROGRAM 'head -n50000 /home/heikki/git-sandbox/ann-benchmarks/items.data';

create index on items_small using hnsw (embedding vector_cosine_ops) WITH (m=16, ef_construction=100);
```
On 'master':
```
postgres=# \timing on
Timing is on.
postgres=# reindex index items_small_embedding_idx ;
REINDEX
Time: 27857.182 ms (00:27.857)
postgres=# reindex index items_small_embedding_idx ;
REINDEX
Time: 27303.268 ms (00:27.303)
postgres=# reindex index items_small_embedding_idx ;
REINDEX
Time: 27510.155 ms (00:27.510)
```
With this PR:
```
postgres=# \timing on
Timing is on.
postgres=# reindex index items_small_embedding_idx ;
REINDEX
Time: 20230.559 ms (00:20.231)
postgres=# reindex index items_small_embedding_idx ;
REINDEX
Time: 19757.048 ms (00:19.757)
postgres=# reindex index items_small_embedding_idx ;
REINDEX
Time: 19648.914 ms (00:19.649)
```