@jkatz 2024-01-05
@hlinnaka Ack. I had tried this test against the previous version and only saw a marginal 1.33% gain. I kept as close to the defaults as possible.

### Setup
* Instance: r7gd.16xlarge
* Storage: NVMe (to isolate any network blips)
* PostgreSQL 16.1
* Vector size: 1536-dim
* Rows: 10,000,000
* Storage: EXTENDED
* Test: Concurrent inserts - 50
* Method:
```sql
INSERT INTO table SELECT generate_random_normalized_vector(VECTOR_DIM)
```
* HNSW build parameters
  * `m = 16`
  * `ef_construction = 64`

(~0.5ms to generate vector)

### Results

master@9a782d29f
```
scaling factor: 1
query mode: simple
number of clients: 50
number of threads: 50
maximum number of tries: 1
number of transactions per client: 200000
number of transactions actually processed: 10000000/10000000
number of failed transactions: 0 (0.000%)
latency average = 57.984 ms
latency stddev = 20.294 ms
initial connection time = 22.598 ms
tps = 861.786527 (without initial connection time)
```

PR@beb1c57073 (prior to rebase)
```
number of clients: 50                                                                                                                                                                                                                                                         
number of threads: 50                                                                                                                                                                                                                                                         
maximum number of tries: 1                                                                                                                                                                                                                                                    
number of transactions per client: 200000                                                                                                                                                                                                                                     
number of transactions actually processed: 10000000/10000000                                                                                                                                                                                                                  
number of failed transactions: 0 (0.000%)                                                                                                                                                                                                                                     
latency average = 57.225 ms                                                                                                                                                                                                                                                   
latency stddev = 19.949 ms                                                                                                                                                                                                                                                    
initial connection time = 22.822 ms                                                                                                                                                                                                                                           
tps = 873.285046 (without initial connection time) 
```

I'll rerun the test with 128-dim and 1536-dim vectors after updating both branches. I'll perform a similar load to yours first to see if I'm seeing the same results.