@ankane 2024-01-25
Hey @hlinnaka, just spent some time on this. I think it breaks `closer` caching (since `e->closer` points to a copy of the neighbor now), which is where the performance gain comes from (as far as I can tell). I also tried a version that makes `c` an array (to avoid creating a list in `HnswUpdateConnection`), but didn't see any performance difference.