@hlinnaka 2024-01-26
> Hey @hlinnaka, just spent some time on this. I think it causes a bug in `closer` caching (since `e` points to a copy of the neighbor now), which is where the performance gain comes from (as far as I can tell). I also tried a version that makes `c` an array (to avoid creating a list in `HnswUpdateConnection`), but didn't see any performance difference.

A-ha, thanks! I'll look into this again.

The interface of SelectNeighbors is super confusing. It modifies the inputs, namely the 'closerSet' flag on the neighbor array and the 'closer' flags on the original HnswCandidate structs. But it it doesn't actually modify the neighbors array itself, to put it in the right order. It requires tight co-operation from the caller. 

I'll think of ways to clear that up. Hopefully to avoid mistakes like this again.